pyproject.toml
[project]
name = "python-google-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "google-adk[database,eval,sqlite,sqlite3]>=0.1.0",
    "litellm>=1.65.5",
    "marimo>=0.13.0",
    "mypy>=1.15.0",
]
